Move Tools/Shared/RuntimeError.h -> Code/Shared/RuntimeError.h

+++ Code/Shared/RuntimeError.h
@@ -1,6 +1,5 @@
 #pragma once
 #include <sstream>
-#include <IOKit/IOKitLib.h>
 
 class RuntimeError : public std::runtime_error {
 public:
@@ -11,7 +10,7 @@ private:
     static std::string fmtMsg(const char* str) {
         char msg[256];
         int sr = snprintf(msg, sizeof(msg), "%s", str);
-        if (sr<0 || sr>=(sizeof(msg)-1)) throw std::runtime_error("failed to create RuntimeError");
+        if (sr<0 || (size_t)sr>=(sizeof(msg)-1)) throw std::runtime_error("failed to create RuntimeError");
         return msg;
     }
     
@@ -19,7 +18,7 @@ private:
     static std::string fmtMsg(const char* fmt, Args&& ...args) {
         char msg[256];
         int sr = snprintf(msg, sizeof(msg), fmt, args...);
-        if (sr<0 || sr>=(sizeof(msg)-1)) throw std::runtime_error("failed to create RuntimeError");
+        if (sr<0 || (size_t)sr>=(sizeof(msg)-1)) throw std::runtime_error("failed to create RuntimeError");
         return msg;
     }
 };
